+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Fixed an issue where high resolution textures weren't read properly from legacy wismt files.
+* Fixed an issue where map textures did not always correctly load the base mip level.
 
 ### Changed
 * Renamed `write_to_file` methods to `save` for all relevant types.